Changes for 5.0.4 version
Switch to Selenium 4 and JDK 11
Implement several improvements:
 - Log should display system property value used during scenario
execution:
	Currently the log file displays only values for properties defined in
files read by ScenarioParametersManager. In certain circumstances it
could be interesting to know values of System properties and Environment
variables as well.
 - Add the possibility to print console output in a separated log file:
	Currently framework is writing full debug information in a log file
(usually debug_YYYYMMDDHHMMSS.log) and simply print output in standard
console. Although the print output is also written in the debug log
file, when running a scenario in a build it's not always easy to get the
console output.
	The purpose of this enhancement is to have the ability to also write
the print output in a separated log file which could be easily got after
a Jenkins or a Travis build run.
	Note that default behavior will be to have this functionality disabled.
Hence its introduction will be silent for users but they can activate it
if necessary in their corresponding builds.
 - Add CloseBrowsers annotation which closes all browsers for test and
step
 - Add sortListAndCompareEquality method to CollectionsUtil
 - Improve SpotTextTimeout.getCondition to use StringUtils.compare
 - Set DigitalizedVersion class public in StringUtils
 - Add filterXmlContent method to StringUtils
 - Print scenario properties usage in console at the end of execution
 - Make some Property and Counters classes and methods public
 - Add and tune some utilities
	+ Add MapUtil class with API to compare maps of strings
	+ Tune doc in CollectionsUtil and StringUtils
	+ Add readPropertiesFile(File) API method in FileUtil
	+ Add IsPartOf comparison to StringUtils
 - Add arrayToList API method on CollectionsUtil
 - Tune Application login when redirection occurred
Fix following issues:
 - ScenarioExecution fails to find the reason when a scenario execution
is stopped
	Sometimes we got following message in console: "For some unknown
reason, while shutting down, ScenarioExecution is not able to retrieve
the reason of the execution stop."
 - Bug in CollectionsUtil.sortListAndCheckEquality method
 - NaN error while displaying scenario counters
 - Tools QA Document title
 - Inaccurate time in message of SpotAbstractTimeout.wait(boolean)
method
 - Workaround for Selenium issue
https://github.com/SeleniumHQ/selenium/issues/11750
 - API key or token stored in properties or environment variables have
their value exposed in log files
 - Bug select when using comparison pattern
 - Lowercase issue while reading tests parameter